FINDING: ACCT-F2026092579 LANE: FINANCIAL (R-178b cast the reused $4 in the fuel expense line insert -- PR 22715 left "inconsistent types deduced for parameter $4")
LANE_CROSS=docs/bus/09-25-2026-LEAD-RULING-R178-FUEL-ENGINE-LINE-AND-NUMBER.md

DOD-A: PASS — active path: createExpenseFromFuelTransaction line 1 and generateExpenseNumber (every expense writer that numbers by load)
DOD-B: N/A — no wizard fields changed
DOD-C: PASS — expense_load_links numbering stays unique; no linkage removed
DOD-D: PASS — economics unchanged: quantity 1 x rate = amount; account and item unchanged
DOD-E: PASS — R-178 dry run gets past both refusals (measured)
VERIFY-1: N/A — no UI
VERIFY-2: PASS — the constraint is satisfied by construction (round(1 x amount) = amount)
VERIFY-3: N/A — no schema change
VERIFY-4: N/A
VERIFY-5: N/A
VERIFY-6: N/A
VERIFY-7: N/A
VERIFY-8: N/A — no security/RLS change

MODULE_PROGRESS: accounting 38 of 40
ITEMS_TOUCHED: ACCT-F2026092578
MIGRATE: N/A

Live=PASS
ROOT CAUSE: PR 22715 reused $4 for both amount_cents (bigint) and rate_cents in one INSERT; Postgres refuses with "inconsistent types deduced for parameter $4", so createExpenseFromFuelTransaction still refused on main. My miss: I merged before the R-178 dry run re-ran on the change.
FIX: $4::bigint at both positions. Guard now fails on the uncast reuse.
GUARD: scripts/verify-fuel-expense-line-and-number-engine.mjs (static + selftest), verify-step 11605.
LIVE PROOF: guard PASS; vitest fuel-expense-document-load-numbering 6/6 and lumper-cash-advance-split 10/10 pass.
REMAINING: R-178 data run (57 fuel rows to their document dates) under AUTH-031 after this merges.

+++ apps/backend/src/fuel/fuel-expense-document.service.ts
@@ -394,7 +394,7 @@ export async function createExpenseFromFuelTransaction(
       )
       -- R-178: an item line must carry quantity/rate/uom with round(quantity*rate_cents)=amount_cents
       -- (expense_lines_item_qty_rate_amount_check). Without them every fuel expense create refused.
-      VALUES ($1::uuid, $2::uuid, 1, $3, $4, $5, $6::uuid, $7, $8::uuid, $9::uuid, 1, $4, 'each')
+      VALUES ($1::uuid, $2::uuid, 1, $3, $4::bigint, $5, $6::uuid, $7, $8::uuid, $9::uuid, 1, $4::bigint, 'each')
     `,
     [
       input.operating_company_id,
